refactor(tasks): reorganize imports in create_event and views modules
