--- strava/tasks/__init__.py
@@ -1,9 +1,0 @@
-from strava.tasks.create_event import create_event
-from strava.tasks.update_activity_weather import update_activity_weather
-from strava.tasks.update_triathlon_score import update_triathlon_score
-
-__all__ = [
-    "create_event",
-    "update_activity_weather",
-    "update_triathlon_score",
-]
